refactor: fix bugs + refactor code
