+++ pages/api/signin.js
@@ -1,8 +1,7 @@
-const { default: axios } = require("axios");
+import { login } from "../calls";
 
 export default (req, res) => {
-  const { query } = req;
-  const { code } = query;
+  const { code } = req.query;
   if (!code) {
     return res.send({
       success: false,
@@ -10,24 +9,9 @@ export default (req, res) => {
     });
   }
 
-  axios
-    .post(
-      `https://github.com/login/oauth/access_token`,
-      {
-        client_id: process.env.NEXT_PUBLIC_GITHUB_CLIENT_ID,
-        client_secret: process.env.NEXT_PUBLIC_GITHUB_CLIENT_SECRET,
-        code,
-      },
-      {
-        headers: {
-          Accept: "application/json",
-        },
-      }
-    )
-    .then((response) => {
-      res.redirect(`/home?accessToken=${response.data.access_token}`);
-    })
-    .catch((err) => {
-      throw err;
-    });
+  login({ code }).then((data) => {
+    res.redirect(`/home?accessToken=${data.access_token}`);
+  }).catch((err) => {
+    throw err;
+  });
 };
